[android] Add shared link view
