+++ toolkit/libtoolkit/src/main/res/values-ru/translatable.xml
@@ -11,4 +11,17 @@
   <string name="sizes_megabytes">Мб</string>
   <string name="sizes_gigabytes">Гб</string>
   <string name="sizes_terabytes">Тб</string>
+
+  <plurals name="days">
+    <item quantity="one">день</item>
+    <item quantity="few">дня</item>
+    <item quantity="other">дней</item>
+  </plurals>
+
+  <plurals name="hours">
+    <item quantity="one">час</item>
+    <item quantity="few">часа</item>
+    <item quantity="other">часов</item>
+  </plurals>
+
 </resources>
