+++ toolkit/libtoolkit/src/main/res/values/translatable.xml
@@ -19,4 +19,15 @@
 
     <string name="export_send_copy">Send a copy</string>
 
+
+    <plurals name="days">
+        <item quantity="one">day</item>
+        <item quantity="other">days</item>
+    </plurals>
+
+    <plurals name="hours">
+        <item quantity="one">hour</item>
+        <item quantity="other">hours</item>
+    </plurals>
+
 </resources>
